# swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Dare To Accept API",
    "version": "1.0.0",
    "description": "API documentation for Dare To Accept platform."
  },
  "host": "localhost:5000",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "paths": {
    "/auth/forgot-password": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Send a password reset email",
        "description": "Allows a user to request a password reset email.",
        "operationId": "forgotPassword",
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/ForgotPasswordRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Password reset email sent"
          },
          "404": {
            "description": "User not found"
          },
          "500": {
            "description": "Error sending email"
          }
        }
      }
    },
    "/auth/login": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Log in a user",
        "description": "Authenticates a user and returns a JWT token.",
        "operationId": "login",
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/LoginRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Login successful"
          },
          "401": {
            "description": "Invalid credentials"
          }
        }
      }
    },
    "/auth/signup": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Sign up a new user",
        "description": "Registers a new user and returns a JWT token.",
        "operationId": "signup",
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/SignupRequest"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "User created successfully"
          },
          "400": {
            "description": "Validation error"
          },
          "500": {
            "description": "Error creating user"
          }
        }
      }
    },
    "/campaigns": {
      "get": {
        "tags": [
          "Campaigns"
        ],
        "summary": "Get all campaigns",
        "description": "Fetches a paginated list of all campaigns.",
        "operationId": "getCampaigns",
        "parameters": [
          {
            "$ref": "#/parameters/PageParam"
          },
          {
            "$ref": "#/parameters/LimitParam"
          }
        ],
        "responses": {
          "200": {
            "description": "List of campaigns",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Campaign"
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Campaigns"
        ],
        "summary": "Create a new campaign",
        "description": "Allows a user to create a new campaign.",
        "operationId": "createCampaign",
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Campaign"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Campaign created"
          },
          "400": {
            "description": "Validation error"
          }
        }
      }
    },
...
    "/comments/{campaignId}": {
      "get": {
        "tags": [
          "Comments"
        ],
        "summary": "Get comments for a campaign",
        "description": "Fetches all comments associated with a specific campaign.",
        "operationId": "getCampaignComments",
        "parameters": [
          {
            "in": "path",
            "name": "campaignId",
            "required": true,
            "type": "string",
            "description": "The ID of the campaign to fetch comments for."
          }
        ],
        "responses": {
          "200": {
            "description": "List of comments for the campaign",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Comment"
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Comments"
        ],
        "summary": "Add a comment to a campaign",
        "description": "Allows a user to add a comment to a specific campaign.",
        "operationId": "addCampaignComment",
        "parameters": [
          {
            "in": "path",
            "name": "campaignId",
            "required": true,
            "type": "string",
            "description": "The ID of the campaign to add a comment to."
          }
        ],
        "responses": {
          "201": {
            "description": "Comment added successfully"
          },
          "404": {
            "description": "Campaign not found"
          }
        }
      }
    },
...
    "/admin/campaigns/{id}/suspend": {
      "put": {
        "tags": [
          "Admin"
        ],
        "summary": "Suspend a campaign",
        "description": "Suspends a specific campaign for administrative purposes.",
        "operationId": "suspendCampaign",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "string",
            "description": "The ID of the campaign to suspend."
          }
        ],
        "responses": {
          "200": {
            "description": "Campaign suspended successfully",
            "schema": {
              "$ref": "#/definitions/Campaign"
            }
          }
        }
      }
    },
...
    "/withdrawals": {
      "post": {
        "tags": [
          "Withdrawals"
        ],
        "summary": "Create a withdrawal request",
        "description": "Allows a user to request a withdrawal to their bank account.",
        "operationId": "createWithdrawal",
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "amount": {
                  "type": "number",
                  "description": "Amount to withdraw"
                },
                "bankAccount": {
                  "type": "string",
                  "description": "Bank account details"
                }
              },
              "required": [
                "amount",
                "bankAccount"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Withdrawal request successful"
          },
          "400": {
            "description": "Invalid request"
          },
          "500": {
            "description": "Error processing withdrawal"
          }
        }
      }
    },
...
    "/challenges/{challengeId}/participants/{participantId}/out": {
      "post": {
        "tags": [
          "Challenges"
        ],
        "summary": "Admit defeat in a challenge",
        "description": "Allows a participant to admit defeat in a specific challenge.",
        "operationId": "admitDefeat",
        "parameters": [
          {
            "in": "path",
            "name": "challengeId",
            "required": true,
            "type": "string",
            "description": "The ID of the challenge."
          },
          {
            "in": "path",
            "name": "participantId",
            "required": true,
            "type": "string",
            "description": "The ID of the participant admitting defeat."
          }
        ],
        "responses": {
          "200": {
            "description": "Defeat admitted successfully",
            "schema": {
              "$ref": "#/definitions/Participant"
            }
          },
          "400": {
            "description": "Cannot admit defeat in current status"
          },
          "404": {
            "description": "Participant not found"
          }
        }
      }
    },
...
    "/upload": {
      "post": {
        "tags": [
          "Media"
        ],
        "summary": "Upload media file",
        "description": "Allows the frontend to upload a media file (image or video) as multipart/form-data. The server uploads the media to Cloudinary and returns the media URL in the response.",
        "operationId": "uploadMedia",
        "consumes": [
          "multipart/form-data"
        ],
        "parameters": [
          {
            "in": "formData",
            "name": "file",
            "required": true,
            "type": "file",
            "description": "The media file to upload (image or video)."
          }
        ],
        "responses": {
          "200": {
            "description": "Media uploaded successfully",
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "example": "File uploaded successfully"
                },
                "secureUrl": {
                  "type": "string",
                  "example": "https://res.cloudinary.com/demo/image/upload/v1234567890/sample.jpg"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request or no file uploaded",
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "example": "No file uploaded"
                }
              }
            }
          },
          "500": {
            "description": "Error uploading media",
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "example": "Error uploading media"
                },
                "error": {
                  "type": "string",
                  "example": "Detailed error message"
                }
              }
            }
          }
        }
      }
    },
...
  },
...
  "parameters": {
    "PageParam": {
      "name": "page",
      "in": "query",
      "required": false,
      "type": "integer",
      "description": "Page number for pagination"
    },
    "LimitParam": {
      "name": "limit",
      "in": "query",
      "required": false,
      "type": "integer",
      "description": "Number of items per page"
    },
...
  },
  "definitions": {
    "ForgotPasswordRequest": {
      "type": "object",
      "properties": {
        "email": {
          "type": "string",
          "format": "email"
        }
      },
      "required": [
        "email"
      ]
    },
    "LoginRequest": {
      "type": "object",
      "properties": {
        "email": {
          "type": "string",
          "format": "email"
        },
        "password": {
          "type": "string",
          "format": "password"
        }
      },
      "required": [
        "email",
        "password"
      ]
    },
    "SignupRequest": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string"
        },
        "email": {
          "type": "string",
          "format": "email"
        },
        "password": {
          "type": "string",
          "format": "password"
        }
      },
      "required": [
        "name",
        "email",
        "password"
      ]
    },
    "Campaign": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string"
        },
        "title": {
          "type": "string"
        },
        "description": {
          "type": "string"
        },
        "goal": {
          "type": "number"
        },
        "status": {
          "type": "string"
        }
      }
    },
    "User": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string"
        },
        "name": {
          "type": "string"
        },
        "email": {
          "type": "string"
        }
      }
    },
    "Comment": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string"
        },
        "text": {
          "type": "string"
        },
        "user": {
          "$ref": "#/definitions/User"
        }
      }
    },
...
  }
}